Transitioning to using arrays/ArrayLists
Currently, the program just prints out the numbers as they're found, instead of storing them in some sort of data structure. The basic layout for using an ArrayList filled with integer arrays has been planned out, with a little bit typed up but most of it planned out in writing (Pictures of the brainstorming is in the folder 'KKC Brainstorm Photos').