"Finished" Print style, moving onto Array style
Made slight adjustments to the print style method so it works correctly except for still printing the incomplete subchains. The Array version is a hot mess right now because I'm thinking of having the ArrayList hold 2D arrays. I very liberally commented out the code for this version, but it is still very unorganized and hard to follow right now. Lots of comments will be erased and added later on.